kbuild: Fix asm-generic/vmlinux.lds.h for LD_DEAD_CODE_DATA_ELIMINATION
KEEP more tables, and add the function/data section wildcard to more
section selections.

This is a little ad-hoc at the moment, but kernel code should be moved
to consistently use .text..x (note: double dots) for explicit sections
and all references to it in the linker script can be made with
TEXT_MAIN, and similarly for other sections.

For now, let's see if major architectures move to enabling this option
then we can do some refactoring passes. Otherwise if it remains unused
or superseded by LTO, this may not be required.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Change-Id: Id6df6118a4a0d55d6ecb91a4fb03ac77b72b2109
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Signed-off-by: TogoFire <italomellopereira@gmail.com>