power: qpnp-fg-gen3: fix soc not keeping full after charging overnight
The issue happened when the temperature was cool after charge_done.
(chip->health was not POWER_SUPPLY_HEALTH_GOOD). In this condition,
chip->charge_full was not set, and then battery level dropped.

Set chip->charge_full to true if chip->health is
POWER_SUPPLY_HEALTH_GOOD or POWER_SUPPLY_HEALTH_COOL
or POWER_SUPPLY_HEALTH_WARM to fix it.

Bug: 122704179
Signed-off-by: Jack Wu <wjack@google.com>
Change-Id: I0e9dd19051118d6281d1fc4989672f07b1371912
Signed-off-by: Lau <laststandrighthere@gmail.com>