sched: Do not use IPIs for remote wakeups if idle
Do not wake up and idle CPU for the chance of a cache hit. This avoids
unnecessary energy consumption. Plus, if we need to wake an idle CPU to
handle a remote wakeup, we will induce additional scheduling latency
since the remote CPU will be unavailable while leaving an idle state.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Change-Id: I7406f3ecfc1351fd8cebd3996bd2d40411e900af