fingerprint: fpc: Allow owner only to read-write to irq sysfs node
DEVICE_ATTR_RO() makes the permission read-only which doesn't retain
the functionality of "0600" permission. Allow the owner only to read
and write to the sysfs node to check the interrupt status of fpc
appropriately and hence use DEVICE_ATTR() macro to set the custom
"0600" permission and thus correct the faulty permission. For this,
bring back the irq_ack() function that was dropped with commit
1c92cc0cffb8b5 ("fingerprint: fpc: Don't measure latency").

This fixes the following userspace log:

06-16 21:09:21.950 E/fpc_tac (1220): fpc_sysfs_node_write write failed 5

Fixes: 441ac6221e1fd ("fingerprint: fpc: Change DEVICE_ATTR() to DEVICE_ATTR_WO|RO()")
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ic224e5076660dabcca58cd6f05459f600814784a
Signed-off-by: TogoFire <togofire@mailfence.com>